+++ lib/puppet/type/shared/timeout.rb
@@ -2,6 +2,6 @@ newparam(:timeout) do
   include EasyType
   include EasyType::Mungers::Integer
 
-  desc "Timeout for this operation"
+  desc 'Timeout for this operation'
 
 end