+++ lib/puppet/type/wls_safagent.rb
@@ -40,6 +40,7 @@ module Puppet
     parameter :domain
     parameter :name
     parameter :safagent_name
+    parameter :timeout
     property :servicetype
     property :persistentstore
     property :persistentstoretype