+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - removed extrapropertiesvalues of wls_datasource, you can use now extraproperties with key=value,key1=value
 - removed extrapropertiesvalues of wls_foreign_server, you can use now extraproperties with key=value,key1=value
 - New Order parameter on the wls_authentication_provider type so you can change the order
+- Timout parameter for all the wls types, can be used to override the default timeout (120s) on every wls resource
 
 ## 1.0.18
 - small fix for the server_template type with the arguments parameter